requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==45.0.7
Django==5.2.5
django-allauth==65.11.0
django-autocomplete-light==3.12.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-multiselectfield==1.0.1
django-parler==2.3
django-tailwind==4.2.0
django-unfold==0.64.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
lxml==6.0.1
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
